refactor(api): split identity slice out of store.ts
Phase 2 wave 4: move IdentityStore's implementation (getUser, handles,
account lifecycle, email prefs) into store/slices/identity.ts, following the
delegation pattern used by every prior wave-4 slice. claimHandle's two
implementations (~120 lines combined) move into their own
identity-claim-handle.ts to keep identity.ts under the 500-line new-file cap.

deleteAccountIdentity is the one exception: it stays implemented directly on
InMemoryStore/FirestoreStore rather than delegating to the identity slice,
since it is the account-erasure orchestrator and already reaches across
nearly every other slice's state via the "documented exception" fields this
wave has been adding. Moving it into InMemoryIdentityStore would relocate
that reach without removing it, at the cost of the slice class depending on
every sibling it doesn't own. Its declaration moves to a small
AccountErasureStore interface declared directly in store.ts, alongside the
comment explaining why.
