+++ eslint-rules/comment-prose-baseline.json
@@ -295,6 +295,8 @@
     "apps/api/src/store/slices/access-tokens.ts": 120,
     "apps/api/src/store/slices/access.ts": 242,
     "apps/api/src/store/slices/agent-keys.ts": 282,
+    "apps/api/src/store/slices/identity-claim-handle.ts": 19,
+    "apps/api/src/store/slices/identity.ts": 419,
     "apps/api/src/store/slices/notifications.ts": 116,
     "apps/api/src/store/slices/oauth.ts": 83,
     "apps/api/src/store/slices/player-data.ts": 916,
